+++ entities/src/bpmn_20_ad_hoc_sub_process.rs
@@ -7,11 +7,11 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "bpmn_20_ad_hoc_sub_process")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     /// SUPER FIELD : SubProcess
-    pub super_sub_process: i32,
+    pub super_sub_process: i64,
     /// COMPLEX FIELD : AdHocSubProcess-completionCondition
-    pub completion_condition: i32,
+    pub completion_condition: i64,
     /// SIMPLE FIELD : AdHocSubProcess-ordering
     pub ordering: AdHocOrdering,
     /// SIMPLE FIELD : AdHocSubProcess-cancelRemainingInstances
