+++ entities/src/bpmn_20_assignment.rs
@@ -7,13 +7,13 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "bpmn_20_assignment")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     /// SUPER FIELD : BaseElement
-    pub super_base_element: i32,
+    pub super_base_element: i64,
     /// COMPLEX FIELD : Assignment-from
-    pub from: i32,
+    pub from: i64,
     /// COMPLEX FIELD : Assignment-to
-    pub to: i32,
+    pub to: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
