+++ entities/src/bpmn_20_collaboration.rs
@@ -7,11 +7,11 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "bpmn_20_collaboration")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     /// SUPER FIELD : RootElement
-    pub super_root_element: i32,
+    pub super_root_element: i64,
     /// COMPLEX FIELD : Collaboration-conversationAssociations
-    pub conversation_associations: i32,
+    pub conversation_associations: i64,
     /// SIMPLE FIELD : Collaboration-name
     pub name: std::string::String,
     /// SIMPLE FIELD : Collaboration-isClosed
