+++ entities/src/bpmn_20_data_association.rs
@@ -7,13 +7,13 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "bpmn_20_data_association")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     /// SUPER FIELD : BaseElement
-    pub super_base_element: i32,
+    pub super_base_element: i64,
     /// COMPLEX FIELD : DataAssociation-transformation
-    pub transformation: Option<i32>,
+    pub transformation: Option<i64>,
     /// COMPLEX FIELD : DataAssociation-targetRef
-    pub target_ref: i32,
+    pub target_ref: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
