+++ entities/src/bpmn_20_data_store.rs
@@ -7,11 +7,11 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "bpmn_20_data_store")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     /// SUPER FIELD : RootElement
-    pub super_root_element: i32,
+    pub super_root_element: i64,
     /// SUPER FIELD : ItemAwareElement
-    pub super_item_aware_element: i32,
+    pub super_item_aware_element: i64,
     /// SIMPLE FIELD : DataStore-name
     pub name: std::string::String,
     /// SIMPLE FIELD : DataStore-capacity
