+++ entities/src/bpmn_20_error.rs
@@ -7,11 +7,11 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "bpmn_20_error")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     /// SUPER FIELD : RootElement
-    pub super_root_element: i32,
+    pub super_root_element: i64,
     /// COMPLEX FIELD : Error-structureRef
-    pub structure_ref: Option<i32>,
+    pub structure_ref: Option<i64>,
     /// SIMPLE FIELD : Error-name
     pub name: std::string::String,
     /// SIMPLE FIELD : Error-errorCode
