+++ entities/src/bpmn_20_event_definition.rs
@@ -7,9 +7,9 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "bpmn_20_event_definition")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     /// SUPER FIELD : RootElement
-    pub super_root_element: i32,
+    pub super_root_element: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
