+++ entities/src/bpmn_20_extension_attribute_definition.rs
@@ -7,9 +7,9 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "bpmn_20_extension_attribute_definition")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     /// COMPLEX FIELD : ExtensionAttributeDefinition-extensionDefinition
-    pub extension_definition: i32,
+    pub extension_definition: i64,
     /// SIMPLE FIELD : ExtensionAttributeDefinition-name
     pub name: std::string::String,
     /// SIMPLE FIELD : ExtensionAttributeDefinition-type
