+++ entities/src/bpmn_20_human_performer.rs
@@ -7,9 +7,9 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "bpmn_20_human_performer")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     /// SUPER FIELD : Performer
-    pub super_performer: i32,
+    pub super_performer: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
